--- a/src/Ocelot/DownstreamRouteFinder/UrlMatcher/IUrlPathToUrlTemplateMatcher.cs
+++ b/src/Ocelot/DownstreamRouteFinder/UrlMatcher/IUrlPathToUrlTemplateMatcher.cs
@@ -1,9 +1,10 @@
 using Ocelot.Responses;
+using Ocelot.Values;
 
 namespace Ocelot.DownstreamRouteFinder.UrlMatcher
 {
     public interface IUrlPathToUrlTemplateMatcher
      {
-        Response<UrlMatch> Match(string upstreamUrlPath, string upstreamQueryString, string upstreamUrlPathTemplate, bool containsQueryString);
+        Response<UrlMatch> Match(string upstreamUrlPath, string upstreamQueryString, UpstreamPathTemplate pathTemplate);
      }
 }