--- a/src/Ocelot/Values/UpstreamPathTemplate.cs
+++ b/src/Ocelot/Values/UpstreamPathTemplate.cs
@@ -1,3 +1,5 @@
+using System.Text.RegularExpressions;
+
 namespace Ocelot.Values
 {
     public class UpstreamPathTemplate
@@ -8,6 +10,9 @@ namespace Ocelot.Values
             Priority = priority;
             ContainsQueryString = containsQueryString;
             OriginalValue = originalValue;
+            Pattern = template == null ? 
+                new Regex("$^", RegexOptions.Compiled | RegexOptions.Singleline) : 
+                new Regex(template, RegexOptions.Compiled | RegexOptions.Singleline);
         }
 
         public string Template { get; }
@@ -17,5 +22,7 @@ namespace Ocelot.Values
         public bool ContainsQueryString { get; }
 
         public string OriginalValue { get; }
+        
+        public Regex Pattern { get; }
     }
 }