This PR includes a change that represents a significant improvement in both memory allocation and Op/s for matching URLs for routing requests. This is achieved by compiling the Regex used to match a route at template construction time instead of on the fly with each request. This allows the regex engine to optimize the pattern matching quite a bit. There was a change to the signature for IUrlPathToUrlTemplateMatcher needed to achieve this, but it is functionally the same. All tests are passing, though many tests had to be modified to use the new signature. Benchmarks attached below.

Fixes / New Feature #

Performance enhancements.

## Proposed Changes

Prior to these changes, 3.29KB is allocated per request and 274K Op/s is achieved.
```

// * Summary *

BenchmarkDotNet=v0.10.14, OS=manjaro 
AMD Ryzen 7 2700U with Radeon Vega Mobile Gfx, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.500
  [Host]     : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT


   Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |      Op/s | Scaled |  Gen 0 | Allocated |
--------- |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|-------:|----------:|
 Baseline | 3.643 us | 0.0096 us | 0.0085 us | 0.0023 us | 3.627 us | 3.636 us | 3.642 us | 3.651 us | 3.655 us | 274,525.0 |   1.00 | 6.4163 |   3.29 KB |
```

With these changes, only 96B is allocated per request and 10,876K Op/s is achieved.

```
// * Summary *

BenchmarkDotNet=v0.10.14, OS=manjaro 
AMD Ryzen 7 2700U with Radeon Vega Mobile Gfx, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.500
  [Host]     : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT


   Method |     Mean |    Error |   StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |         Op/s | Scaled |  Gen 0 | Allocated |
--------- |---------:|---------:|---------:|----------:|---------:|---------:|---------:|---------:|---------:|-------------:|-------:|-------:|----------:|
 Baseline | 91.94 ns | 1.862 ns | 4.773 ns | 0.5439 ns | 83.93 ns | 89.62 ns | 91.32 ns | 93.40 ns | 114.8 ns | 10,876,595.7 |   1.00 | 0.1826 |      96 B |
```