@margaale 2018-12-22
Great numbers, but what happens when the configuration is changed (and reload) after ocelot has been started?