@philproctor 2018-12-22
It recreates the templates on configuration change. All the properties on the Template class were already effectively read only, so new template regex's are compiled every time they are recreated. The regex is tied to the existing constructor that was already used and that signature has not changed.