@margaale 2018-12-22
Great. Fine by me. It Seems I'm gonna be the first one to approve a PR. It feels kinda awkward.